+++ lib/src/clojure_lsp/feature/file_management.clj
@@ -8,7 +8,6 @@
    [clojure-lsp.feature.diagnostics :as f.diagnostic]
    [clojure-lsp.feature.rename :as f.rename]
    [clojure-lsp.kondo :as lsp.kondo]
-   [clojure-lsp.logger :as logger]
    [clojure-lsp.producer :as producer]
    [clojure-lsp.queries :as q]
    [clojure-lsp.settings :as settings]
@@ -142,12 +141,11 @@
 (defn ^:private notify-references [uri db-before db-after {:keys [db* producer] :as components}]
   (async/thread
     (shared/logging-task
-      :notify-references
+      :internal/notify-references
       (let [uris (shared/logging-task
                    :reference-files/find
                    (reference-uris uri db-before db-after))]
         (when (seq uris)
-          (logger/debug "Analyzing references for files:" uris)
           (shared/logging-task
             :reference-files/analyze
             ;; TODO: We process the dependent and dependency files together, but
@@ -170,7 +168,7 @@
             ;; https://github.com/clojure-lsp/clojure-lsp/issues/1027 and
             ;; https://github.com/clojure-lsp/clojure-lsp/issues/1028.
             (analyze-reference-uris! uris db*))
-          (f.diagnostic/publish-all-diagnostics! uris components)
+          (f.diagnostic/publish-all-diagnostics! uris true components)
           (producer/refresh-code-lens producer))))))
 
 (defn ^:private offsets [lines line character end-line end-character]
@@ -250,12 +248,12 @@
   (loop [state-db @db*]
     (when (>= version (get-in state-db [:documents uri :v] -1))
       (let [kondo-result* (future
-                            (shared/logging-time
-                              (str "changes analyzed by clj-kondo took %s")
+                            (shared/logging-task
+                              :internal/uri-analyzed-by-clj-kondo
                               (lsp.kondo/run-kondo-on-text! text uri db*)))
             depend-result* (future
-                             (shared/logging-time
-                               (str "changes analyzed by clj-depend took %s")
+                             (shared/logging-task
+                               :internal/uri-analyzed-by-clj-depend
                                (lsp.depend/analyze-uri! uri state-db)))
             kondo-result @kondo-result*
             depend-result @depend-result*
@@ -288,11 +286,9 @@
                        ;; we check if file still exists/should be linted
                        (filter #(get-in @db* [:documents %]))
                        (map shared/uri->filename))
-        result (shared/logging-time
-                 "Watched files analyzed, took %s"
-                 (lsp.kondo/run-kondo-on-paths! filenames db* {:external? false} nil))]
+        result (lsp.kondo/run-kondo-on-paths! filenames db* {:external? false} nil)]
     (swap! db* lsp.kondo/db-with-results result)
-    (f.diagnostic/publish-all-diagnostics! uris components)
+    (f.diagnostic/publish-all-diagnostics! uris true components)
     (producer/refresh-test-tree producer uris)
     (doseq [uri uris]
       (when (get-in @db* [:documents uri :v])
@@ -342,7 +338,7 @@
       (async/>!! watched-files-chan created-or-changed))
     (when (seq deleted)
       (shared/logging-task
-        :delete-watched-files
+        :internal/delete-watched-files
         (files-deleted components deleted)))))
 
 (defn did-close [uri {:keys [db*] :as components}]
@@ -379,7 +375,7 @@
                                  (keep (fn [[uri document]] (when (:v document) uri)))
                                  set)]
         (analyze-reference-uris! all-opened-uris db*)
-        (f.diagnostic/publish-all-diagnostics! all-opened-uris components)
+        (f.diagnostic/publish-all-diagnostics! all-opened-uris true components)
         (producer/refresh-code-lens producer)))))
 
 (defn did-save [uri {:keys [db*] :as components}]