+++ lib/src/clojure_lsp/feature/resolve_macro.clj
@@ -13,6 +13,8 @@
 
 (set! *warn-on-reflection* true)
 
+(def logger-tag "[resolve-macro]")
+
 (def ^:private excluded-macros
   '{clojure.core *
     clojure.core.async *
@@ -70,7 +72,7 @@
   (if-let [result (apply producer/show-message-request producer message)]
     result
     (do
-      (logger/error error)
+      (logger/error logger-tag error)
       nil)))
 
 (defn resolve-macro-as!
@@ -95,8 +97,8 @@
                                                   :version (:v document)
                                                   :text (:text document)}
                                                  components)
-              (logger/info (format "Resolving macro as %s. Saving setting into %s" resolved-full-symbol-str kondo-config-path)))
+              (logger/info logger-tag (format "Resolving macro as %s. Saving setting into %s" resolved-full-symbol-str kondo-config-path)))
             (do
-              (logger/error (format "Could not resolve macro at cursor to be resolved as '%s' for path '%s'" resolved-full-symbol-str kondo-config-path))
+              (logger/error logger-tag (format "Could not resolve macro at cursor to be resolved as '%s' for path '%s'" resolved-full-symbol-str kondo-config-path))
               (producer/show-message producer (format "No macro was found at cursor to resolve as '%s'." resolved-full-symbol-str) :error nil))))))
     {:no-op? true}))